requirements.txt
Flask>=2.0.0
Telethon>=1.25.0
APScheduler>=3.9.0
python-dotenv>=0.19.0
cryptography>=38.0.0
aiosqlite>=0.17.0
gunicorn>=20.1.0
Werkzeug>=2.0.0
